+++ packages/bbob-parser/src/index.js
@@ -1,2 +1,2 @@
-export { default, parse } from './parse';
 export { TagNode } from '@bbob/plugin-helper';
+export { default, parse } from './parse';
