+++ packages/bbob-plugin-helper/package.json
@@ -53,5 +53,15 @@
     "lib",
     "src",
     "es"
-  ]
+  ],
+  "exports": {
+    ".": {
+      "types": "./src/index.d.ts",
+      "browser": "./dist/index.min.js",
+      "umd": "./lib/index.js",
+      "import": "./es/index.js",
+      "require": "./lib/index.js"
+    },
+    "./package.json": "./package.json"
+  }
 }
